teman-edu-mvp/requirements.txt
streamlit==1.42.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.11
alembic==1.14.1
bcrypt==4.2.1
pandas==2.2.3
reportlab==4.2.5
python-dotenv==1.0.1
pytest==8.3.4
